requirements.txt

rembg==2.0.72
onnxruntime==1.24.1
pillow==12.1.0

pystray==0.19.5
easygui==0.98.3